+++ app/Services/StudentPortal/ProfileChangeService.php
@@ -107,8 +107,8 @@ class ProfileChangeService
             $schoolId = $filters['school_id'] ?? null;
             $studentId = $filters['student_id'] ?? null;
             $status = $filters['status'] ?? null;
-            $limit = $filters['limit'] ?? 50;
-            $offset = $filters['offset'] ?? 0;
+            $limit = (int) ($filters['limit'] ?? 50);
+            $offset = (int) ($filters['offset'] ?? 0);
 
             if (!$schoolId) {
                 return [
