+++ app/Services/StudentPortal/SpecialRequestService.php
@@ -121,8 +121,8 @@ class SpecialRequestService
         try {
             $schoolId = $filters['school_id'] ?? null;
             $studentId = $filters['student_id'] ?? null;
-            $limit = $filters['limit'] ?? 50;
-            $offset = $filters['offset'] ?? 0;
+            $limit = (int) ($filters['limit'] ?? 50);
+            $offset = (int) ($filters['offset'] ?? 0);
 
             if (!$schoolId) {
                 return [
